requirements.txt
requests==2.18.4
tqdm==4.19.5
beautifulsoup4==4.6.0
PyYAML==3.12
